refactor: update context handling to use gincontext package and adjust golangci.yaml settings
