+++ pkg/gincontext/gin_context_test.go
@@ -1,4 +1,4 @@
-package context
+package gincontext
 
 import (
 	"errors"
